DBDT31647 【临床数据中心3期】ETL管理-作业管理调用ETLhttp执行、生命周期年龄字段类型调整

See merge request lyra4j/shucang/services4j/shujuzx!456

+++ mediinfo-grus-shujuzx/src/main/java/cn/mediinfo/grus/shujuzx/dto/shengmingzqpzs/SC_ZD_ShengMingZQDto.java
@@ -3,6 +3,7 @@ package cn.mediinfo.grus.shujuzx.dto.shengmingzqpzs;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import lombok.Data;
+import org.apache.poi.hpsf.Decimal;
 
 @Data
 public class SC_ZD_ShengMingZQDto {
@@ -13,9 +14,9 @@ public class SC_ZD_ShengMingZQDto {
     @Schema(description = "生命周期名称")
     private String shengMingZQMC;
     @Schema(description = "起始年龄")
-    private Integer qiShiNL;
+    private Decimal qiShiNL;
     @Schema(description = "结束年龄")
-    private Integer jieShuNL;
+    private Decimal jieShuNL;
     @Schema(description = "男性图片")
     private String nanXingTP;
     @Schema(description = "女性图片")
